Introduce invite code step into onboarding
Re-arrange order of steps, still roughing out the details
