+++ xcode/Subconscious/Shared/Components/AppView.swift
@@ -242,6 +242,7 @@ enum AppAction {
     case submitFirstRunProfileStep
     case submitFirstRunSphereStep
     case submitFirstRunRecoveryStep
+    case submitFirstRunInviteStep
     case submitFirstRunDoneStep
     
     case requestOfflineMode
@@ -461,9 +462,10 @@ enum AppDatabaseState {
 }
 
 enum FirstRunStep {
-    case profile
     case sphere
     case recovery
+    case profile
+    case invite
     case done
 }
 
@@ -677,6 +679,11 @@ struct AppModel: ModelProtocol {
             )
         case .submitFirstRunRecoveryStep:
             return submitFirstRunRecoveryStep(
+                state: state,
+                environment: environment
+            )
+        case .submitFirstRunInviteStep:
+            return submitFirstRunInviteStep(
                 state: state,
                 environment: environment
             )
@@ -1491,7 +1498,7 @@ struct AppModel: ModelProtocol {
         return update(
             state: state,
             actions: [
-                .pushFirstRunStep(.profile)
+                .pushFirstRunStep(.sphere)
             ],
             environment: environment
         )
@@ -1510,7 +1517,7 @@ struct AppModel: ModelProtocol {
             state: state,
             actions: [
                 .submitNickname(nickname),
-                .pushFirstRunStep(.sphere)
+                .pushFirstRunStep(.invite)
             ],
             environment: environment
         )
@@ -1523,6 +1530,7 @@ struct AppModel: ModelProtocol {
         return update(
             state: state,
             actions: [
+                .createSphere,
                 .pushFirstRunStep(.recovery)
             ],
             environment: environment
@@ -1530,6 +1538,19 @@ struct AppModel: ModelProtocol {
     }
     
     static func submitFirstRunRecoveryStep(
+        state: AppModel,
+        environment: AppEnvironment
+    ) -> Update<AppModel> {
+        return update(
+            state: state,
+            actions: [
+                .pushFirstRunStep(.profile)
+            ],
+            environment: environment
+        )
+    }
+    
+    static func submitFirstRunInviteStep(
         state: AppModel,
         environment: AppEnvironment
     ) -> Update<AppModel> {
