+++ xcode/Subconscious/Shared/Components/FirstRun/FirstRunView.swift
@@ -7,44 +7,9 @@
 import ObservableStore
 import SwiftUI
 
-struct InviteCodeRedeemedView: View {
-    var gatewayId: String
-    
-    var body: some View {
-        VStack(spacing: AppTheme.unit2) {
-            HStack(spacing: AppTheme.unit2) {
-                Image(systemName: "checkmark.circle")
-                    .resizable()
-                    .frame(width: 16, height: 16)
-                    .opacity(0.5)
-                Text("Invitation accepted")
-            }
-            .font(.body)
-            .bold()
-            
-            HStack(spacing: AppTheme.unit) {
-                Text("Gateway ID")
-                    .bold()
-                Text(gatewayId)
-            }
-            .font(.caption)
-        }
-        .foregroundColor(.secondary)
-    }
-}
-
 struct FirstRunView: View {
     @ObservedObject var app: Store<AppModel>
     @Environment(\.colorScheme) var colorScheme
-    
-    var inviteCodeCaption: String {
-        switch app.state.inviteCodeRedemptionStatus {
-        case .failed(_):
-            return String(localized: "Could not redeem invite code")
-        case _:
-            return String(localized: "You can find your invite code in your welcome email")
-        }
-    }
     
     var body: some View {
         NavigationStack(
@@ -57,96 +22,35 @@ struct FirstRunView: View {
             VStack(spacing: AppTheme.padding) {
                 Spacer()
                 
-                if !app.state.inviteCodeFormField.hasFocus {
-                    StackedGlowingImage() {
-                        Image("sub_logo").resizable()
-                    }
-                    .aspectRatio(contentMode: .fit)
-                    .frame(
-                        minWidth: 32,
-                        maxWidth: AppTheme.onboarding.heroIconSize,
-                        minHeight: 32,
-                        maxHeight: AppTheme.onboarding.heroIconSize
-                    )
-                    
-                    Spacer()
+                StackedGlowingImage() {
+                    Image("sub_logo").resizable()
                 }
+                .aspectRatio(contentMode: .fit)
+                .frame(
+                    minWidth: 32,
+                    maxWidth: AppTheme.onboarding.heroIconSize,
+                    minHeight: 32,
+                    maxHeight: AppTheme.onboarding.heroIconSize
+                )
                 
                 Text("Subconscious is a place to garden thoughts and share them with others.")
                     .foregroundColor(.secondary)
                     .font(.callout)
                     .multilineTextAlignment(.center)
                     
-                Text("It’s powered by Noosphere, a decentralized protocol, so your data belongs to you.")
+                Text("It's a personal notebook designed to help you discover connections and think new thoughts.")
                     .foregroundColor(.secondary)
                     .font(.callout)
                     .multilineTextAlignment(.center)
                 
                 Spacer()
                 
-                if let gatewayId = app.state.gatewayId {
-                    InviteCodeRedeemedView(
-                        gatewayId: gatewayId
-                    )
-                } else {
-                    ValidatedFormField(
-                        alignment: .center,
-                        placeholder: "Enter your invite code",
-                        field: app.state.inviteCodeFormField,
-                        send: Address.forward(
-                            send: app.send,
-                            tag: AppAction.inviteCodeFormField
-                        ),
-                        caption: inviteCodeCaption,
-                        onFocusChanged: { focused in
-                            // User finished editing the field
-                            if !focused {
-                                app.send(.submitInviteCodeForm)
-                            }
-                        }
-                    )
-                    .textFieldStyle(.roundedBorder)
-                    .textInputAutocapitalization(.never)
-                    .disableAutocorrection(true)
-                    .disabled(app.state.gatewayOperationInProgress)
-                }
-                
-                if !app.state.inviteCodeFormField.hasFocus {
-                    Spacer()
-                    
-                    Button(action: {
-                        app.send(.submitFirstRunWelcomeStep)
-                    }, label: {
-                        Text("Get Started")
-                    })
-                    .buttonStyle(PillButtonStyle())
-                    .disabled(app.state.gatewayId == nil)
-                }
-                
-                // MARK: Use Offline
-                VStack {
-                    if app.state.gatewayId == .none {
-                        HStack(spacing: AppTheme.unit) {
-                            Text("No invite code?")
-                                .font(.caption)
-                                .foregroundColor(.secondary)
-                            
-                            Button(action: {
-                                app.send(.requestOfflineMode)
-                            }, label: {
-                                Text("Use offline")
-                                    .font(.caption)
-                            })
-                        }
-                    }
-                }.padding(
-                    .init(
-                        top: 0,
-                        leading: 0,
-                        bottom: AppTheme.tightPadding,
-                        trailing: 0
-                    )
-                )
+                Button(action: {
+                    app.send(.submitFirstRunWelcomeStep)
+                }, label: {
+                    Text("Get Started")
+                })
+                .buttonStyle(PillButtonStyle())
             }
             .navigationTitle("Welcome to Subconscious")
             .navigationBarTitleDisplayMode(.inline)
@@ -165,12 +69,11 @@ struct FirstRunView: View {
                     FirstRunSphereView(app: app)
                 case .recovery:
                     FirstRunRecoveryView(app: app)
+                case .invite:
+                    FirstRunInviteView(app: app)
                 case .done:
                     FirstRunDoneView(app: app)
                 }
-            }
-            .onAppear {
-                app.send(.createSphere)
             }
         }
     }
